"more refactoring; updates #832"
