+++ software/utils/debug/colors.py
@@ -1,0 +1,11 @@
+
+class Colors:
+    ENDC = '\033[m'
+    GRAY = '\033[90m'
+    RED = '\033[91m'
+    GREEN = '\033[92m'
+    ORANGE = '\033[93m'
+    BLUE = '\033[94m'
+    PURPLE = '\033[95m'
+    YELLOW = '\033[96m'
+    WHITE = '\033[97m'
